STR-000: Replace ChatwootWidget with ZohoWidget and remove Chatwoot component
